product tile

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import connectToDb from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/product";
 import { NextResponse } from "next/server";
 
@@ -7,28 +8,37 @@ export const dynamic = "force-dynamic";
 export async function DELETE(req) {
   try {
     await connectToDb();
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
+    const isAuthUser = await AuthUser(req);
+    if (isAuthUser?.role === "admin") {
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get("id");
 
-    if (!id) {
-      return NextResponse.json({
-        success: true,
-        message: "Product id is required ",
-      });
-    }
+      if (!id) {
+        return NextResponse.json({
+          success: true,
+          message: "Product id is required ",
+        });
+      }
 
-    const deleteProduct = await Product.findByIdAndDelete(id);
+      const deleteProduct = await Product.findByIdAndDelete(id);
 
-    if (deleteProduct) {
-      return NextResponse.json({
-        success: true,
-        message: "product deleted successfully !! ",
-      });
-    } else {
+      if (deleteProduct) {
+        return NextResponse.json({
+          success: true,
+          message: "product deleted successfully !! ",
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: "failed to delete",
+        });
+      }
+    }else{
       return NextResponse.json({
         success: false,
-        message: "failed to delete",
+        message: "Unauthorized",
       });
+
     }
   } catch (e) {
     return NextResponse.json({
